ci: green the pipeline (LoC split + clippy + fmt)
Make the full PR #62 merge chain pass CI. Three classes of fix:

1. LoC ≤500 (file-size-cap): split scheduler/helpers.rs (834 → 476) by
   moving its `#[cfg(test)] mod thinking_loop_tests` (360 lines) to
   helpers_tests.rs via `#[path]` — logical child of `helpers`, so
   `use super::*` resolves exactly as before; zero production change.
   (decode_logits_step.rs / serve.rs are on the CI allow_list — left.)

2. clippy (deny clippy::all) — fixes, several PRE-EXISTING on the base
   branch and unmasked once upstream crates compiled clean:
   - atlas-core config/methods.rs: manual checked division → checked_div
   - atlas-kernels build_codegen.rs: generated consts emit `&str`/`&[u8]`
     not `&'static …` (clears 15 redundant_static_lifetimes in the
     generated target_ptx.rs; fn-return `'static` kept — not linted)
   - spark-runtime buffers/sizes.rs + spark-server preflight.rs: manual
     checked division → checked_div().map().unwrap_or()
   - spark-model forward_layers.rs + spark-server phase_promote_prefills.rs:
     descending sort_by → sort_(unstable_)by_key(|x| Reverse(..))
   - spark-server openai/annotations.rs: loop+let-else-break → while let
   - spark-server scheduler/helpers.rs: iter().any(|t| t==X) → contains(&X)

3. cargo fmt --all (workspace) — normalizes the clippy edits plus
   pre-existing base fmt debt (mtp_head/new.rs, qwen35_dense.rs,
   decode_logits_seq/step.rs).

Verified in atlas-builder: fmt --check 0 diffs, no non-allowlisted .rs
>500, `cargo clippy --tests --workspace --keep-going` zero errors,
`cargo test --workspace` all green (incl. 342 spark-server + 23
helpers). Behavior-preserving throughout (test-mod move, semantically
identical clippy rewrites, whitespace-only fmt).
