+++ crates/spark-model/src/layers/mtp_head/new.rs
@@ -77,107 +77,128 @@ impl MtpHead {
             (None, None, None)
         } else {
             match quant {
-            MtpQuantization::Nvfp4 => {
-                let gate_nvfp4 = quantize_to_nvfp4(
-                    &weights.moe_gate,
-                    config.num_experts,
-                    h,
-                    gpu,
-                    absmax_k,
-                    nvfp4_k,
-                    stream,
-                )?;
-                let mut experts = Vec::with_capacity(weights.experts.len());
-                for (i, de) in weights.experts.iter().enumerate() {
-                    let gate_proj =
-                        quantize_to_nvfp4(&de.gate_proj, inter, h, gpu, absmax_k, nvfp4_k, stream)?;
-                    let up_proj =
-                        quantize_to_nvfp4(&de.up_proj, inter, h, gpu, absmax_k, nvfp4_k, stream)?;
-                    let down_proj =
-                        quantize_to_nvfp4(&de.down_proj, h, inter, gpu, absmax_k, nvfp4_k, stream)?;
-                    experts.push(crate::weight_map::ExpertWeight {
-                        gate_proj,
-                        up_proj,
-                        down_proj,
-                    });
-                    if (i + 1) % 128 == 0 {
-                        tracing::info!(
-                            "  MTP experts quantized: {}/{}",
-                            i + 1,
-                            weights.experts.len()
-                        );
+                MtpQuantization::Nvfp4 => {
+                    let gate_nvfp4 = quantize_to_nvfp4(
+                        &weights.moe_gate,
+                        config.num_experts,
+                        h,
+                        gpu,
+                        absmax_k,
+                        nvfp4_k,
+                        stream,
+                    )?;
+                    let mut experts = Vec::with_capacity(weights.experts.len());
+                    for (i, de) in weights.experts.iter().enumerate() {
+                        let gate_proj = quantize_to_nvfp4(
+                            &de.gate_proj,
+                            inter,
+                            h,
+                            gpu,
+                            absmax_k,
+                            nvfp4_k,
+                            stream,
+                        )?;
+                        let up_proj = quantize_to_nvfp4(
+                            &de.up_proj,
+                            inter,
+                            h,
+                            gpu,
+                            absmax_k,
+                            nvfp4_k,
+                            stream,
+                        )?;
+                        let down_proj = quantize_to_nvfp4(
+                            &de.down_proj,
+                            h,
+                            inter,
+                            gpu,
+                            absmax_k,
+                            nvfp4_k,
+                            stream,
+                        )?;
+                        experts.push(crate::weight_map::ExpertWeight {
+                            gate_proj,
+                            up_proj,
+                            down_proj,
+                        });
+                        if (i + 1) % 128 == 0 {
+                            tracing::info!(
+                                "  MTP experts quantized: {}/{}",
+                                i + 1,
+                                weights.experts.len()
+                            );
+                        }
                     }
+                    let shared_gate = quantize_to_nvfp4(
+                        &weights.shared_expert.gate_proj,
+                        inter,
+                        h,
+                        gpu,
+                        absmax_k,
+                        nvfp4_k,
+                        stream,
+                    )?;
+                    let shared_up = quantize_to_nvfp4(
+                        &weights.shared_expert.up_proj,
+                        inter,
+                        h,
+                        gpu,
+                        absmax_k,
+                        nvfp4_k,
+                        stream,
+                    )?;
+                    let shared_down = quantize_to_nvfp4(
+                        &weights.shared_expert.down_proj,
+                        h,
+                        inter,
+                        gpu,
+                        absmax_k,
+                        nvfp4_k,
+                        stream,
+                    )?;
+                    let moe_weights = MoeWeights {
+                        gate: weights.moe_gate,
+                        shared_expert: crate::weight_map::ExpertWeight {
+                            gate_proj: shared_gate,
+                            up_proj: shared_up,
+                            down_proj: shared_down,
+                        },
+                        shared_expert_gate: weights.shared_expert_gate,
+                        experts,
+                        router_pre_norm: None,
+                        correction_bias: None,
+                    };
+                    let moe = MoeLayer::new(
+                        moe_weights,
+                        config.num_experts,
+                        Some(gate_nvfp4),
+                        gpu,
+                        config,
+                    )?;
+                    (Some(moe), None, None)
                 }
-                let shared_gate = quantize_to_nvfp4(
-                    &weights.shared_expert.gate_proj,
-                    inter,
-                    h,
-                    gpu,
-                    absmax_k,
-                    nvfp4_k,
-                    stream,
-                )?;
-                let shared_up = quantize_to_nvfp4(
-                    &weights.shared_expert.up_proj,
-                    inter,
-                    h,
-                    gpu,
-                    absmax_k,
-                    nvfp4_k,
-                    stream,
-                )?;
-                let shared_down = quantize_to_nvfp4(
-                    &weights.shared_expert.down_proj,
-                    h,
-                    inter,
-                    gpu,
-                    absmax_k,
-                    nvfp4_k,
-                    stream,
-                )?;
-                let moe_weights = MoeWeights {
-                    gate: weights.moe_gate,
-                    shared_expert: crate::weight_map::ExpertWeight {
-                        gate_proj: shared_gate,
-                        up_proj: shared_up,
-                        down_proj: shared_down,
-                    },
-                    shared_expert_gate: weights.shared_expert_gate,
-                    experts,
-                    router_pre_norm: None,
-                    correction_bias: None,
-                };
-                let moe = MoeLayer::new(
-                    moe_weights,
-                    config.num_experts,
-                    Some(gate_nvfp4),
-                    gpu,
-                    config,
-                )?;
-                (Some(moe), None, None)
-            }
-            MtpQuantization::Fp8 | MtpQuantization::Bf16 => {
-                let mut experts_g = Vec::with_capacity(weights.experts.len());
-                for (i, de) in weights.experts.iter().enumerate() {
-                    let gate_proj = q(&de.gate_proj, inter, h)?;
-                    let up_proj = q(&de.up_proj, inter, h)?;
-                    let down_proj = q(&de.down_proj, h, inter)?;
-                    experts_g.push((gate_proj, up_proj, down_proj));
-                    if (i + 1) % 128 == 0 {
-                        tracing::info!(
-                            "  MTP experts quantized: {}/{}",
-                            i + 1,
-                            weights.experts.len()
-                        );
+                MtpQuantization::Fp8 | MtpQuantization::Bf16 => {
+                    let mut experts_g = Vec::with_capacity(weights.experts.len());
+                    for (i, de) in weights.experts.iter().enumerate() {
+                        let gate_proj = q(&de.gate_proj, inter, h)?;
+                        let up_proj = q(&de.up_proj, inter, h)?;
+                        let down_proj = q(&de.down_proj, h, inter)?;
+                        experts_g.push((gate_proj, up_proj, down_proj));
+                        if (i + 1) % 128 == 0 {
+                            tracing::info!(
+                                "  MTP experts quantized: {}/{}",
+                                i + 1,
+                                weights.experts.len()
+                            );
+                        }
                     }
+                    let shared = (
+                        q(&weights.shared_expert.gate_proj, inter, h)?,
+                        q(&weights.shared_expert.up_proj, inter, h)?,
+                        q(&weights.shared_expert.down_proj, h, inter)?,
+                    );
+                    (None, Some(experts_g), Some(shared))
                 }
-                let shared = (
-                    q(&weights.shared_expert.gate_proj, inter, h)?,
-                    q(&weights.shared_expert.up_proj, inter, h)?,
-                    q(&weights.shared_expert.down_proj, h, inter)?,
-                );
-                (None, Some(experts_g), Some(shared))
-            }
             }
         };
 
@@ -243,7 +264,11 @@ impl MtpHead {
             h2 = h * 2,
             qd = nq * hd * 2,
             ffn = ffn_kind,
-            ne = if dense_ffn_generic.is_some() { 0 } else { config.num_experts },
+            ne = if dense_ffn_generic.is_some() {
+                0
+            } else {
+                config.num_experts
+            },
             ev = effective_vocab,
             fv = config.vocab_size,
             lm = (effective_vocab * h / 2) as f64 / (1024.0 * 1024.0),
